Confirm Dialog

+++ trilhas/src/app/componentes/crud/crud.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { CrudService } from '../../services/crud.service';
 import { HomeService } from '../../services/home.service';
+import { DialogconfirmService } from '../../services/dialogconfirm.service';
 import { ActivatedRoute, Router } from '@angular/router';
 
 @Component({
@@ -10,13 +11,13 @@ import { ActivatedRoute, Router } from '@angular/router';
 })
 export class CrudComponent implements OnInit {  
 
-  usuario: any = {};
-  logo: any = {};
+  usuario: any = [];
+  logo: any = [];
   alert: boolean = false;
   alertD: boolean = false;
 
   constructor(private crudServ: CrudService, private homeServ: HomeService,
-    private route: ActivatedRoute,
+    private route: ActivatedRoute, private dialogConfirmServ: DialogconfirmService,
     private router: Router) {
     crudServ.getUsuario().subscribe(usuario => {
       this.usuario = usuario['allUsers'];
@@ -35,13 +36,20 @@ export class CrudComponent implements OnInit {
 
   onDelete(item){
     this.usuario.splice(item-1,1)
-    this.crudServ.onDelete(item).subscribe((result) => {
-      console.warn("result",result)
-      this.crudServ.getUsuario().subscribe(usuario => {
-        this.usuario = usuario['allUsers'];
-        console.log(this.usuario);
-      });
+    this.dialogConfirmServ.confirm("Deseja excluir esse usuário?").then((podeDeletar:boolean) => {
+      if (podeDeletar) {
+        this.crudServ.onDelete(item).subscribe((result) => {
+          console.warn("result",result)
+          this.crudServ.getUsuario().subscribe(usuario => {
+            this.usuario = usuario['allUsers'];
+            console.log(this.usuario);
+          });
+        })
+      } else{
+        return false;
+      }
     })
+    
   }
 
   closeAlert(){
